[CK2] Horse Lord Invasion [Playable, Unfinished Features]
A small Mini-Mod for Crusader Kings II where a Horse invades Central Europe in a similar fashion to the Aztecs and Mongols. Requires Horse Lords if you want to play as the horse (VIA Console).

Currently unfinished, but its completly playable. I wanted to add a few variants to the Horse so its not always Slepinir, make its religion dynamic so its not always Catholic, and allow it to spawn in more places than a few very specific provinces.

USAGE: If anyone wishes to finish my work, shoot me a message here or on steam and please credit me in the mod and as a creator of the mod. :)